✅ fix store tests
